extract some logic to *.api.* files
